+++ draw.py
@@ -230,7 +230,7 @@ def text_draw_callback_px(self, context):
 
     font_id = 0
     text_body = getattr(self, 'text_body', 'Running')
-    text_size = getattr(self, 'text_size', 25.0)
+    text_size = getattr(self, 'text_size', 20.0)
     text_color = getattr(self, 'text_color', (0.0, 0.5, 1.0)) # red -> (0.8, 0.1, 0.2)
     text_position = getattr(self, 'text_position', (120, 120))
 
@@ -239,12 +239,24 @@ def text_draw_callback_px(self, context):
     ## optional : Center X position accordinga to text dimension (may offset a lot on x when text is updating)
     # dimensions = blf.dimensions(font_id, text_body)
     # text_position = (text_position[0] - (dimensions[0] / 2), text_position[1])
-    blf.position(font_id, *text_position, 0) # Leave out z at 0
     
     ## Shadow (no error but do not see anything)
     # blf.shadow(font_id, 5, 0.0, 0.0, 0.0, 1.0) # fontid, blur level (0, 3, 5) or outline (6)), r, g, b, a
     # blf.shadow_offset(fontid, 4, -4) # fontid, x, y
-    blf.draw(font_id, text_body)
+    
+    if '\n' in text_body:
+        ## display line by line
+        for i, line in enumerate(text_body.split('\n')):
+            ## Offset each line by the height of the previous
+            width, height = blf.dimensions(font_id, line)
+            ## a good ratio is 50% of the height
+            offset_down = i * (height + height / 2) # line + spacing
+            blf.position(font_id, text_position[0], text_position[1] - offset_down, 0)
+            blf.draw(font_id, line)
+    else:
+        ## Standard display
+        blf.position(font_id, *text_position, 0) # Leave out z at 0
+        blf.draw(font_id, text_body)
 
 
 def gp_plane_callback(self, context):